refactor: move BRE service to business logic
