+++ SeaPublicWebsite.BusinessLogic/ExternalServices/Bre/Enums/BreFloorType.cs
@@ -1,4 +1,4 @@
-﻿namespace SeaPublicWebsite.ExternalServices.Models;
+﻿namespace SeaPublicWebsite.BusinessLogic.ExternalServices.Bre.Enums;
 
 public enum BreFloorType
 {
