+++ SeaPublicWebsite.BusinessLogic/ExternalServices/Bre/Enums/BreGlazingType.cs
@@ -1,4 +1,4 @@
-﻿namespace SeaPublicWebsite.ExternalServices.Models
+﻿namespace SeaPublicWebsite.BusinessLogic.ExternalServices.Bre.Enums
 {
     public enum BreGlazingType
     {
